# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sync/sessions:
    post:
      operationId: createSyncSession
      tags: [sync]
      summary: Create a resumable two-node sync session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          description: Sync session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}:
    get:
      operationId: getSyncSession
      tags: [sync]
      summary: Read a sync session and its package reports
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Sync session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}/handshake:
    post:
      operationId: handshakeSyncSession
      tags: [sync]
      summary: Exchange vector clocks
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          description: Handshake result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}/manifest:
    post:
      operationId: prepareSyncManifest
      tags: [sync]
      summary: Prepare a delta manifest from a vector clock
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          description: Delta manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}/packages:
    post:
      operationId: applySyncPackage
      tags: [sync]
      summary: Validate and apply a received package atomically
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          description: Sync application report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}/packages/{packageId}:
    get:
      operationId: getSyncPackage
      tags: [sync]
      summary: Read a saved sync package report
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
        - { name: packageId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Sync package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/sessions/{sessionId}/ack:
    post:
      operationId: acknowledgeSyncPackage
      tags: [sync]
      summary: Acknowledge a received sync package
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          description: Acknowledgement result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/pairings:
    post:
      operationId: createSyncPairing
      tags: [sync]
      summary: Create a short-lived single-use pairing code
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          description: Pairing code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/pairings/consume:
    post:
      operationId: consumeSyncPairing
      tags: [sync]
      summary: Consume a pairing code and trust a peer node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          description: Trusted node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
...
  /sync/trusted-nodes/{nodeId}/revoke:
    post:
      operationId: revokeTrustedSyncNode
      tags: [sync]
      summary: Revoke a trusted node without deleting audit history
      parameters:
        - { name: nodeId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Revoked node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/conflicts:
    get:
      operationId: listSyncConflicts
      tags: [sync]
      summary: List open, resolved, or deferred sync conflicts
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, resolved, deferred, all] } }
      responses:
        "200":
          description: Conflict queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JsonObject"
  /sync/conflicts/{id}/resolve:
    post:
      operationId: resolveSyncConflict
      tags: [sync]
      summary: Approve, reject, or defer a conflict
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          description: Conflict decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/relay/packages:
    get:
      operationId: listRelayPackages
      tags: [sync]
      summary: List opaque encrypted relay packages
      parameters:
        - { name: sessionId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Relay package metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JsonObject"
    post:
      operationId: uploadRelayPackage
      tags: [sync]
      summary: Store an opaque encrypted relay package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          description: Relay package metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /sync/relay/packages/{relayId}:
    get:
      operationId: downloadRelayPackage
      tags: [sync]
      summary: Download opaque relay bytes when requested
      parameters:
        - { name: relayId, in: path, required: true, schema: { type: string } }
        - { name: download, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: Relay package metadata or encrypted bytes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
...
components:
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...